Refactor Types by moving to Schema files
Moves most of the types to schema files to consolidate functionality.

* Move describe-dataset types to describe-dataset schema
* Move fetch-dataset-geography types to fetch-dataset-geography schema
