--- types/fetch-dataset-geography.types.ts
@@ -1,7 +1,0 @@
-import { z } from "zod";
-
-import {
-	FetchDatasetGeographyInputSchema
-} from '../schema/dataset-geography.schema.js'
-
-export type FetchDatasetGeographyArgs = z.infer<typeof FetchDatasetGeographyInputSchema>;
